Add custom caching to geocoding api responses
The API doesn't give the response header "Expires" or "max-age"
attribute, which means the responses are not automatically cached.
This commit creates copy of the original response, gives it
"Expires" attribute and manually caches them. Then when new request is
fired, the script first checks if there is no cached response, or if
the cached response is expired. If so, brand new request is sent to
the API. Otherwise, if there is cached response that is not expired,
it retrieves that cached response.